requirements.txt
requests~=2.32.3
beautifulsoup4~=4.12.3
lxml~=5.3.0
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1
psycopg2~=2.9.10